ShellPkg: acpiview: SLIT: Validate System Locality count
1. Test if the SlitSystemLocalityCount pointer has been sucessfully
updated before it is later used to control the SLIT table parsing
logic in acpiview.

2. Check if the 'Number of System Localities' provided can be
represented in the SLIT table. The table 'Length' field is a 32-bit
value while the 'Number of System Localities' field is 64-bit long.

3. Check if the SLIT matrix fits in the table buffer. If N is the SLIT
locality count, then the matrix used to represent the localities is
N*N bytes long. The ACPI table length must be big enough to fit the
matrix.

4. Remove (now) redundant 64x64 bit multiplication.

Signed-off-by: Krzysztof Koch <krzysztof.koch@arm.com>